fix css files not recognized: added typing files
I added typing.d.ts within source which should work to ensure that any build command and syntax highlighter will recognize .css files as modules which can be added to our files. This should fix the bug of build not working and it giving an error on highlighting that it cannot find the file.
Also fixed small routing bug.