refactor: Use the api client instead of react query for requests.
We can now use the backend API client instead of react query. This gives better type-checking and notifies us of errors when the API gets updated.

This also means we no longer need to define our types and services in the frontend since these are provided by the package.

We have also removed the dashboard since the backend does not support this yet on main.